fix: modified husky data ingestion endpoints
