+++ apps/web-api/src/internals/members.controller.ts
@@ -1,10 +1,11 @@
-import { Controller, UseGuards, Body, BadRequestException, NotFoundException } from '@nestjs/common';
+import { Controller, UseGuards, Body, BadRequestException, NotFoundException, Param } from '@nestjs/common';
 import { Api, initNestServer, ApiDecorator } from '@ts-rest/nest';
 import { apiInternals } from 'libs/contracts/src/lib/contract-internals';
-import { InternalUpdateMemberDto, ResponseMemberSchema } from 'libs/contracts/src/schema';
+import { InternalUpdateMemberDto, ResponseMemberSchema, ResponseMemberWithRelationsSchema } from 'libs/contracts/src/schema';
 import { ApiOkResponseFromZod } from '../decorators/api-response-from-zod';
 import { InternalAuthGuard } from '../guards/auth.guard';
 import { MembersService } from '../members/members.service';
+import { InternalsService } from './internals.service';
 
 const server = initNestServer(apiInternals);
 type RouteShape = typeof server.routeShapes;
@@ -13,7 +14,8 @@ type RouteShape = typeof server.routeShapes;
 @UseGuards(InternalAuthGuard)
 export class MembersController {
   constructor(
-    private readonly membersService: MembersService
+    private readonly membersService: MembersService,
+    private readonly internalsService: InternalsService
   ) {}
 
   @Api(server.route.updateTelagramUid)
@@ -31,5 +33,11 @@ export class MembersController {
     } else {
         throw new BadRequestException('Telegram handle cannot be empty');
     }    
+  }
+
+  @Api(server.route.getMemberDetails)
+  @ApiOkResponseFromZod(ResponseMemberWithRelationsSchema)
+  async getMemberDetails(@Param('uid') uid: string) {
+    return await this.internalsService.getMemberDetails(uid);
   }
 }
