+++ apps/web-api/src/internals/pl-events.controller.ts
@@ -3,13 +3,14 @@ import { ApiParam } from '@nestjs/swagger';
 import { Request } from 'express';
 import { Api, initNestServer, ApiDecorator } from '@ts-rest/nest';
 import { apiInternals } from 'libs/contracts/src/lib/contract-internals';
-import { ResponsePLEventGuestSchemaWithRelationsSchema } from 'libs/contracts/src/schema';
+import { ResponsePLEventGuestSchemaWithRelationsSchema, ResponsePLEventSchemaWithRelationsSchema } from 'libs/contracts/src/schema';
 import { ApiOkResponseFromZod } from '../decorators/api-response-from-zod';
 import { PLEventGuestsService } from '../pl-events/pl-event-guests.service';
 import { PLEventsService } from '../pl-events/pl-events.service';
 import { PrismaQueryBuilder } from '../utils/prisma-query-builder';
 import { prismaQueryableFieldsFromZod } from '../utils/prisma-queryable-fields-from-zod';
 import { InternalAuthGuard } from '../guards/auth.guard';
+import { InternalsService } from './internals.service';
 
 const server = initNestServer(apiInternals);
 type RouteShape = typeof server.routeShapes;
@@ -19,7 +20,8 @@ type RouteShape = typeof server.routeShapes;
 export class PLEventsInternalController {
   constructor(
     private readonly eventGuestsService: PLEventGuestsService, 
-    private readonly eventService: PLEventsService  
+    private readonly eventService: PLEventsService,
+    private readonly internalsService: InternalsService
   ) {}
 
   @Api(server.route.getPLEventGuestsByLocation)
@@ -43,5 +45,11 @@ export class PLEventsInternalController {
       };
     }
     return await this.eventGuestsService.getPLEventGuestsByLocation(locationUid, builtQuery);
+  }
+
+  @Api(server.route.getIrlEventDetails)
+  @ApiOkResponseFromZod(ResponsePLEventSchemaWithRelationsSchema)
+  async getIrlEventDetails(@Param('uid') uid: string) {
+    return await this.internalsService.getIrlEventDetails(uid);
   }
 }
